[temp] +transformer.to(accelerator.device) p2

+++ examples/stable-diffusion/training/train_text_to_image_sd3.py
@@ -1156,7 +1156,7 @@ def main(args):
             text_encoder_one.gradient_checkpointing_enable()
             text_encoder_two.gradient_checkpointing_enable()
             text_encoder_three.gradient_checkpointing_enable()
-    transformer.to(accelerator.device)
+    transformer.to(accelerator.device, dtype=weight_dtype)
 
     def unwrap_model(model):
         model = accelerator.unwrap_model(model)
